refactor and move tests into its own folder
